// src/pepr/operator/crd/generated/istio/virtualservice-v1beta1.ts
export class VirtualService extends GenericKind {
  /**
   * Configuration affecting label/content routing, sni routing, etc. See more details at:
   * https://istio.io/docs/reference/config/networking/virtual-service.html
   */
  spec?: Spec;
  status?: { [key: string]: unknown };
}

/**
 * Configuration affecting label/content routing, sni routing, etc. See more details at:
 * https://istio.io/docs/reference/config/networking/virtual-service.html
 */
export interface Spec {
  /**
   * A list of namespaces to which this virtual service is exported.
   */
  exportTo?: string[];
  /**
   * The names of gateways and sidecars that should apply these routes.
   */
  gateways?: string[];
  /**
   * The destination hosts to which traffic is being sent.
   */
  hosts?: string[];
  /**
   * An ordered list of route rules for HTTP traffic.
   */
  http?: HTTP[];
  /**
   * An ordered list of route rules for opaque TCP traffic.
   */
  tcp?: TCP[];
  /**
   * An ordered list of route rule for non-terminated TLS & HTTPS traffic.
   */
  tls?: Tl[];
}

export interface HTTP {
  /**
   * Cross-Origin Resource Sharing policy (CORS).
   */
  corsPolicy?: CorsPolicy;
  /**
   * Delegate is used to specify the particular VirtualService which can be used to define
   * delegate HTTPRoute.
   */
  delegate?: Delegate;
  /**
   * A HTTP rule can either return a direct_response, redirect or forward (default) traffic.
   */
  directResponse?: DirectResponse;
  /**
   * Fault injection policy to apply on HTTP traffic at the client side.
   */
  fault?: Fault;
  headers?: HTTPHeaders;
  /**
   * Match conditions to be satisfied for the rule to be activated.
   */
  match?: HTTPMatch[];
  /**
   * Mirror HTTP traffic to a another destination in addition to forwarding the requests to
   * the intended destination.
   */
  mirror?: PurpleMirror;
  mirror_percent?: number;
  mirrorPercent?: number;
  /**
   * Percentage of the traffic to be mirrored by the `mirror` field.
   */
  mirrorPercentage?: MirrorPercentageObject;
  /**
   * Specifies the destinations to mirror HTTP traffic in addition to the original destination.
   */
  mirrors?: MirrorElement[];
  /**
   * The name assigned to the route for debugging purposes.
   */
  name?: string;
  /**
   * A HTTP rule can either return a direct_response, redirect or forward (default) traffic.
   */
  redirect?: Redirect;
  /**
   * Retry policy for HTTP requests.
   */
  retries?: Retries;
  /**
   * Rewrite HTTP URIs and Authority headers.
   */
  rewrite?: Rewrite;
  /**
   * A HTTP rule can either return a direct_response, redirect or forward (default) traffic.
   */
  route?: HTTPRoute[];
  /**
   * Timeout for HTTP requests, default is disabled.
   */
  timeout?: string;
}
// ...
export interface TCP {
  /**
   * Match conditions to be satisfied for the rule to be activated.
   */
  match?: TCPMatch[];
  /**
   * The destination to which the connection should be forwarded to.
   */
  route?: TCPRoute[];
}
// ...
export interface Tl {
  /**
   * Match conditions to be satisfied for the rule to be activated.
   */
  match?: TlMatch[];
  /**
   * The destination to which the connection should be forwarded to.
   */
  route?: TlRoute[];
}